requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
psycopg[binary,pool]==3.2.1
